+++ composer.json
@@ -18,7 +18,7 @@
   "prefer-stable": true,
   "autoload": {
     "psr-4": {
-      "SuiteCRM\\api\\": "api"
+      "": "lib/"
     },
     "psr-0": {
       "": "src/"
