+++ lib/SuiteCRM/api/.htaccess
@@ -3,5 +3,5 @@ RewriteEngine On
 RewriteCond %{REQUEST_FILENAME} !-f
 RewriteCond %{REQUEST_FILENAME} !-d
 RewriteRule ^(.*?)$ public/index.php/$1 [L]
-RewriteRule .* - [env=HTTP_AUTHORIZATION:%{HTTP:Authorization}]
+#RewriteRule .* - [env=HTTP_AUTHORIZATION:%{HTTP:Authorization}]
