+++ modules/Administration/UpgradeAccess.php
@@ -65,6 +65,8 @@ RedirectMatch 403 {$ignoreCase}/+files\.md5\$
     RewriteBase {$basePath}
     RewriteRule ^cache/jsLanguage/(.._..).js$ index.php?entryPoint=jslang&module=app_strings&lang=$1 [L,QSA]
     RewriteRule ^cache/jsLanguage/(\w*)/(.._..).js$ index.php?entryPoint=jslang&module=$1&lang=$2 [L,QSA]
+    RewriteRule ^api/(.*?)$ lib/SuiteCRM/api/public/index.php/$1 [L]
+    RewriteRule ^api/(.*)$ - [env=HTTP_AUTHORIZATION:%{HTTP:Authorization}]
 </IfModule>
 # END SUGARCRM RESTRICTIONS
 EOQ;
